pantalla usuarios

+++ src/main/java/ar/edu/davinci/adoptame/repository/UsuarioRepository.java
@@ -10,9 +10,10 @@ import org.springframework.data.repository.query.Param;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called personaRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
 
 
     public Usuario buscarUsuarioByEmail(@Param("email")String email);
+
 
 }
